requirements.txt
fastapi
uvicorn
pillow
pytesseract
python-multipart
requests
python-dotenv
